feat(reactRouter): add a react router
- create a components folder
- create a routes folder
- add home, login and error components
- add home, login, and error views
- route the the pages

[Finishes #162949032]
