+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "node-sass": "^4.11.0",
     "prop-types": "^15.6.2",
+    "fork-ts-checker-webpack-plugin": "^0.5.2",
     "react": "^16.8.0",
     "react-dom": "^16.8.0",
     "react-redux": "^6.0.0",
@@ -13,7 +14,8 @@
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
     "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.85.0"
+    "semantic-ui-react": "^0.85.0",
+    "typescript": "^3.3.3"
   },
   "scripts": {
     "start": "react-scripts start",
