+++ src/components/Error.js
@@ -1,14 +1,13 @@
-import React from "react"
-
+import React from 'react';
 
 const Error = () =>{
-    return(
-        <div>
-            <h1>
-                Page does not exist!!!
-            </h1>
-        </div>
-    );
+  return(
+    <div>
+      <h1>
+        Page does not exist!!!
+      </h1>
+    </div>
+  );
 };
 
 export default Error;
