+++ src/components/Home.js
@@ -1,12 +1,18 @@
-import React from "react"
-import  HomeView from "../views/HomeView";
+import React from 'react';
+import logo from '../ahlogo.png';
+import '../App.scss';
 
-class Home extends React.Component{
-    render(){
-        return(
-            <HomeView/>
-        );
-    }
-}
+const HomeView = () =>
+  (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+            Welcome to Author's Haven
+        </p>
+            The place of choice.
+      </header>
+    </div>
+  );
 
-export default Home;
+export default HomeView;
