--- src/containers/app/App.test.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import App from './index';
-
-describe('RootComponent', () => {
-  it('app should render without crashing', () => {
-    const wrapper = shallow(<App />);
-    expect(wrapper.length).toBe(1);
-  });
-});
-
-it('renders without crashing', () => {});
