--- src/containers/app/state/action.js
@@ -1,5 +1,0 @@
-import { COUNT_UPDATE } from './types';
-
-export const clickAction = () => ({
-  type: COUNT_UPDATE,
-});
