--- src/containers/app/state/types.js
@@ -1,2 +1,0 @@
-export const COUNT_UPDATE = 'COUNT_UPDATE';
-export const COUNT_UPDATE_SUCCESS = 'COUNT_UPDATE_SUCCESS';
