+++ src/index.js
@@ -2,15 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import './index.scss';
-import './index.css';
-import 'semantic-ui-css/semantic.min.css'
+import 'semantic-ui-css/semantic.min.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import configureStore from './store';
-import App from './containers/app';
+import configureStore from './redux/store';
 
 const store = configureStore();
-
 ReactDOM.render(
   <Provider store={store}>
     <App />
