+++ src/redux/sagas/sagas.js
@@ -1,0 +1,3 @@
+function* updateCountSaga() {}
+
+export default updateCountSaga;
