+++ src/redux/store/rootReducer.js
@@ -1,0 +1,5 @@
+import { combineReducers } from 'redux';
+
+const rootReducer = combineReducers({});
+
+export default rootReducer;
