+++ src/views/ErrorView.js
@@ -1,13 +1,11 @@
-import React from "react"
+import React from 'react';
 
-const Error = () =>{
-    return(
-        <div>
-            <h1>
-                Page does not exist!!!
-            </h1>
-        </div>
-    );
-};
+const Error = () => (
+  <div>
+    <h1>
+        Page does not exist!!!
+    </h1>
+  </div>
+);
 
 export default Error;
