+++ src/views/HomeView.js
@@ -1,20 +1,12 @@
-import React from "react"
-import logo from '../logo.svg';
-import '../App.css';
+import React from 'react';
+import HomeView from '../components/Home';
 
+class Home extends React.Component {
+  render() {
+    return (
+      <HomeView />
+    );
+  }
+}
 
-
-const HomeView = () =>
-(
-    <div className="App">
-        <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-            Welcome to Author's Haven
-        </p>
-            The place of choice.
-        </header>
-    </div>
-);
-
-export default HomeView;
+export default Home;
